# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Podoru is a Docker container management platform similar to EasyPanel.\nIt provides APIs for managing containers, services, networks, and Docker Swarm clusters.",
        "title": "Podoru API",
        "contact": {
            "name": "Podoru Team",
            "url": "https://github.com/podoru/spinner-podoru"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/teams": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all teams the current user is a member of",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "List teams",
                "responses": {
                    "200": {
                        "description": "List of teams",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_podoru_spinner-podoru_internal_domain_entity.TeamWithRole"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new team. The creator becomes the owner.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Create team",
                "parameters": [
                    {
                        "description": "Team data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_internal_domain_entity.TeamCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Team created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_internal_domain_entity.Team"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "409": {
                        "description": "Slug already exists",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/teams/{teamId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get team details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Get team",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Team details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_internal_domain_entity.Team"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid team ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "Not a team member",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Team not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update team details. Requires admin or owner role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Update team",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Team update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_internal_domain_entity.TeamUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated team",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_internal_domain_entity.Team"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "Not a team member or insufficient permissions",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Team not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a team. Only the owner can delete a team.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "Delete team",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Team deleted"
                    },
                    "400": {
                        "description": "Invalid team ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "Not team owner",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Team not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_podoru_spinner-podoru_internal_domain_entity.Team": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_podoru_spinner-podoru_internal_domain_entity.TeamCreate": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "slug": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
...
        "github_com_podoru_spinner-podoru_internal_domain_entity.TeamRole": {
            "type": "string",
            "enum": [
                "owner",
                "admin",
                "member"
            ],
            "x-enum-varnames": [
                "TeamRoleOwner",
                "TeamRoleAdmin",
                "TeamRoleMember"
            ]
        },
        "github_com_podoru_spinner-podoru_internal_domain_entity.TeamUpdate": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
        "github_com_podoru_spinner-podoru_internal_domain_entity.TeamWithRole": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_podoru_spinner-podoru_internal_domain_entity.TeamRole"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_podoru_spinner-podoru_pkg_response.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_podoru_spinner-podoru_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_podoru_spinner-podoru_pkg_response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.ErrorInfo"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_podoru_spinner-podoru_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}